restrict deleteing post

+++ App/controllers/ListingController.php
@@ -2,7 +2,9 @@
 
 namespace App\Controllers;
 
+use Framework\Authorization;
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
 
 class ListingController
@@ -22,7 +24,7 @@ class ListingController
    */
   public function index()
   {
-    $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
     loadView('listings/index', [
       'listings' => $listings
@@ -76,7 +78,7 @@ class ListingController
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     $newListingData = array_map('sanitize', $newListingData);
 
@@ -138,9 +140,16 @@ class ListingController
     ];
     $listings = $this->db->query('SELECT * FROM listiings WHERE id = :id', $params)->fetch();
 
+    // Check if listing exists
     if (!$listings) {
       ErrorController::notFound('Listing not found!');
       return;
+    }
+
+    //  Authorization
+    if (!Authorization::isOwner($listings->user_id)) {
+      $_SESSION['error_message'] = 'You are not autorized to delete this listing';
+      return redirect('/listing/' . $listings->$id);
     }
 
     $this->db->query('DELETE FROM listings WHERE id= :id', $params);
